Ready refactor, squashed
Ready Refactor

* ExecEntry will now run ready checks for Spells, AA and Discs, instead of just spells.
* * This should remove mountains of checks, and for AA, mountains of *duplicate* checks.

* Refactor SpellReady, AAReady, DiscReady
* * Now all three will use a new function CastCheck() for uniformity and code brevity.

* Add SongReady
* * Songs will have the same level of checks for mana/end/etc, and will use SpellReady (which checks gem timer).

* Refactor Targeted... functions
* * Entry conditions should now simply use TargetCheck(), which is only needed to check LOS and ensure we aren't targeting a corpse or dead target, rather than having duplicate code and checks.

Update 1

* For simplification, baked range check into AbilityReady.
* * May revisit this, justification is to try to keep configs simple.

* Moved TargetCheck to targeting

* Added helper for checking if entry target is MA.
* Updated various class configs to assist in playtesting.

* Removed new target check function
* * Upon reflection, removed function. I think the checks are unnecessary overhead. We would only be calling this on hostile actions.

We were checking that our targets were alive, and that they were in LOS.

If they aren't alive, it is a momentary until the MA picks a better target. We will just fail a very small amount of time later, and it is more likely to result in less total calls IMO.

If they aren't in LOS, we need to move anyway; so we should just let our event handlers do our thing. If not, we could theoretically just be standing there for a full rotation check and never fixing the LOS issue to begin with.

* Rebased main

* Baked gem timer check into rotation check

Rather than having a gem timer check in almost every condition with longer recast, tied it to bAllowMem. If we are in combat, we will not try to use a spell unless the gem is available. out of combat, we will bypass this check (so buffs can process/etc). Spells/Buffs with very long refresh times (e.g, NDT)  can still use a GemReady as an entry condition.

* Adjusted ready check with SelfBuffAACheck
No need to do all those stacking checks and whatnot if the abil isn't available. Early out.